LCCI-Deploy/lcci-site/backend-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.20.0
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
email-validator==2.1.0
httpx==0.25.2

